Move signals functionality to models.py in support. (#2616)
